# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/auth/login":{"post":{"operationId":"login","summary":"Login User","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Successfully logged in user!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"tags":["auth"]}},"/auth/user":{"get":{"operationId":"getCurrentUser","summary":"Get Current User","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully get current user details!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CurrentUserResponse"}}}}},"tags":["auth"]}},"/auth/forget-password":{"post":{"operationId":"forgetPassword","summary":"Forget Password user!","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgetPasswordRequestDto"}}}},"responses":{"200":{"description":"OTP sent successfully!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgetPasswordResponseDto"}}}}},"tags":["auth"]}},"/user":{"post":{"operationId":"registerUser","summary":"Register User","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"tags":["user"]}},"/master/currency":{"get":{"operationId":"fetchCurrencies","summary":"Fetch currencies!","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully fetched currencies!","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CurrenciesResponseDto"}}}}}},"tags":["master"]}},"/master/invoice-types":{"get":{"operationId":"fetchInvoiceTypes","summary":"Fetch invoice types","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully fetched invoice-types!","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InvoiceTypesResponseDto"}}}}}},"tags":["master"]}},"/master/gst-types":{"get":{"operationId":"fetchGstTypes","summary":"Fetch gst types","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully fetched gst-types!","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GstTypesResponseDto"}}}}}},"tags":["master"]}},"/master/invoice-status":{"get":{"operationId":"fetchInvoiceStatues","summary":"Fetch invoice statues","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully fetched invoice-statues!","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InvoiceStatusesResponseDto"}}}}}},"tags":["master"]}},"/invoice":{"get":{"operationId":"fetchInvoices","summary":"Fetch all invoices!","parameters":[{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}},{"name":"invoiceStatusId","required":false,"in":"query","example":"0874a355-495a-4a17-a54e-1cc87ad714b0","schema":{"type":"string"}},{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"responses":{"200":{"description":"List of Invoices!","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PaginatedDto"},{"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/InvoiceResponseDto"}},"meta":{"type":"any","default":{"totalItems":2,"itemCount":2,"itemsPerPage":2,"totalPages":1,"currentPage":1}}}}]}}}}},"tags":["invoice"]},"post":{"operationId":"submitAInvoice","summary":"Submit a Invoice!","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitInvoiceRequestDto"}}}},"responses":{"201":{"description":"Successfully submitted a invoice!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvoiceResponseDto"}}}}},"tags":["invoice"]}},"/invoice/{id}":{"get":{"operationId":"fetchInvoiceById","summary":"Fetch Invoice by id!","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully fetched invoice!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvoiceResponseDto"}}}}},"tags":["invoice"]}},"/vendor":{"get":{"operationId":"fetchVendors","summary":"Fetch all vendors!","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully fetched all vendors!","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VendorResponseDto"}}}}}},"tags":["vendor"]},"post":{"operationId":"createVendor","summary":"Create a vendor!","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVendorRequestDto"}}}},"responses":{"201":{"description":"Successfully created a vendor!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VendorResponseDto"}}}}},"tags":["vendor"]}},"/vendor/{id}":{"get":{"operationId":"fetchVendorById","summary":"Fetch vendor by id!","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully fetched vendor!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VendorResponseDto"}}}}},"tags":["vendor"]},"put":{"operationId":"updateVendor","summary":"Update vendor!","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateVendorRequestDto"}}}},"responses":{"200":{"description":"Vendor updated successfully!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VendorResponseDto"}}}}},"tags":["vendor"]},"delete":{"operationId":"deleteVendor","summary":"Delete vendor!","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"responses":{"default":{"description":"Vendor deleted successfully!"}},"tags":["vendor"]}},"/user-access-status/{id}":{"get":{"operationId":"fetchAccessStatues","summary":"Fetch access statuses","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully fetched access statues!","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserAccessStatusResponseDto"}}}}}},"tags":["user-access"]}},"/upload":{"post":{"operationId":"fileUpload","summary":"Get AWS pre-signed URL","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadRequest"}}}},"responses":{"200":{"description":"AWS pre-signed URL generated successfully!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadResponse"}}}}},"tags":["File Upload & Delete"]}},"/upload/delete":{"delete":{"operationId":"fileDelete","summary":"Remove file from AWS","parameters":[{"name":"Authorization","in":"header","required":false,"description":"bearer token","allowEmptyValue":true,"example":"Bearer <token>","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadRequest"}}}},"responses":{"200":{"description":"File removed successfully!"}},"tags":["File Upload & Delete"]}}},"info":{"title":"RIL Invoice Module","description":"API for RIL Invoice Module built with NestJS","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"JWTBearerAuth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","description":"Enter a JWT token to authorize the requests..."}},"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"User Email!","example":"john@email.com"},"password":{"type":"string","description":"User password!","example":"Test@1234"}},"required":["email","password"]},"LoginResponseDto":{"type":"object","properties":{"access_token":{"type":"string","readOnly":true,"example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImNhOTZmMzNlLWY2NzktNGExOC04OGIwLTZkODkyY2NhNThlOSIsInVzZXJfcm9sZV9jb2RlIjoiViIsImlhdCI6MTcxNjkxMTA3MywiZXhwIjoxNzE2OTk3NDczfQ.UyJK6RyesamU-bXjhGknWXWPOvtDN3bvUoDlnXzEMtE"}},"required":["access_token"]},"CurrentUserResponse":{"type":"object","properties":{"id":{"type":"string","readOnly":true,"example":"ca96f33e-f679-4a18-88b0-6d892cca58e9"},"user_role_code":{"type":"string","readOnly":true,"example":"ca96f33e-f679-4a18-88b0-6d892cca58e9"}},"required":["id","user_role_code"]},"ForgetPasswordRequestDto":{"type":"object","properties":{"email":{"type":"string","description":"Enter email!","example":"aritra.sen@contenterra.com"},"newPassword":{"type":"string","description":"Enter new password!","example":"Test@12345"},"otp":{"type":"number","description":"Enter the OTP sent to your email!","example":"123456"}},"required":["email"]},"ForgetPasswordResponseDto":{"type":"object","properties":{"message":{"type":"string","readOnly":true,"example":"User password updated successfully!"}},"required":["message"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","description":"User Email!","example":"john@email.com"},"password":{"type":"string","description":"User password!","example":"Test@1234"},"user_role_id":{"type":"string","description":"vendor user_role_id","example":"b652b2ca-5732-410c-ac5b-7e157846ad67"}},"required":["email","password","user_role_id"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"string","readOnly":true,"example":"cb9f2c82-2910-4766-a634-54502dfb01e1","format":"v4"},"created_at":{"format":"date-time","type":"string","readOnly":true},"updated_at":{"format":"date-time","type":"string","readOnly":true},"is_active":{"type":"boolean","readOnly":true},"email":{"type":"string","readOnly":true,"example":"test@email.com"},"password":{"type":"string","readOnly":true,"example":"$2a$12$4vOrcrWE6L42OmEC2/y5WuBGg/sCHPjFMBvLdfO5kAsL5lLAFJV3u"},"is_verified":{"type":"boolean","readOnly":true,"example":false},"user_role_des":{"type":"string","readOnly":true,"example":"VENDOR"},"user_role_code":{"type":"string","readOnly":true,"example":"V"},"otp":{"type":"number","readOnly":true,"example":1234},"otp_creation_dt":{"format":"date-time","type":"string","readOnly":true}},"required":["id","created_at","updated_at","is_active","email","password","is_verified","user_role_des","user_role_code","otp","otp_creation_dt"]},"CurrenciesResponseDto":{"type":"object","properties":{"id":{"type":"string","readOnly":true,"example":"cb9f2c82-2910-4766-a634-54502dfb01e1","format":"v4"},"created_at":{"format":"date-time","type":"string","readOnly":true},"updated_at":{"format":"date-time","type":"string","readOnly":true},"is_active":{"type":"boolean","readOnly":true},"currency_code":{"type":"string","readOnly":true,"example":"INR"},"currency_des":{"type":"string","readOnly":true,"example":"INDIAN RUPEES"},"country_code":{"type":"string","readOnly":true,"example":"IND"},"country_name":{"type":"string","readOnly":true,"example":"INDIA"}},"required":["id","created_at","updated_at","is_active","currency_code","currency_des","country_code","country_name"]},"InvoiceTypesResponseDto":{"type":"object","properties":{"id":{"type":"string","readOnly":true,"example":"cb9f2c82-2910-4766-a634-54502dfb01e1","format":"v4"},"created_at":{"format":"date-time","type":"string","readOnly":true},"updated_at":{"format":"date-time","type":"string","readOnly":true},"is_active":{"type":"boolean","readOnly":true},"invoice_type_des":{"type":"string","readOnly":true,"example":"TAX"},"invoice_type_code":{"type":"string","readOnly":true,"example":"TAX"}},"required":["id","created_at","updated_at","is_active","invoice_type_des","invoice_type_code"]},"GstTypesResponseDto":{"type":"object","properties":{"id":{"type":"string","readOnly":true,"example":"cb9f2c82-2910-4766-a634-54502dfb01e1","format":"v4"},"created_at":{"format":"date-time","type":"string","readOnly":true},"updated_at":{"format":"date-time","type":"string","readOnly":true},"is_active":{"type":"boolean","readOnly":true},"gst_type_code":{"type":"string","readOnly":true,"example":"CGST"}},"required":["id","created_at","updated_at","is_active","gst_type_code"]},"InvoiceStatusesResponseDto":{"type":"object","properties":{"id":{"type":"string","readOnly":true,"example":"cb9f2c82-2910-4766-a634-54502dfb01e1","format":"v4"},"created_at":{"format":"date-time","type":"string","readOnly":true},"updated_at":{"format":"date-time","type":"string","readOnly":true},"is_active":{"type":"boolean","readOnly":true},"invoice_sts_code":{"type":"string","readOnly":true,"example":"STS"},"invoice_sts_des":{"type":"string","readOnly":true,"example":"ST SUBMITTED"}},"required":["id","created_at","updated_at","is_active","invoice_sts_code","invoice_sts_des"]},"PaginatedDto":{"type":"object","properties":{}},"AttachmentDto":{"type":"object","properties":{"attachment_type_id":{"type":"string","description":"attachment_type_id","example":"8a403b87-b64d-4648-a79d-1fbb62c081bb"},"file_name":{"type":"string","description":"file_name","example":"filename"},"file_path":{"type":"string","description":"file_path","example":"www.example.com"}},"required":["attachment_type_id","file_name","file_path"]},"SubmitInvoiceRequestDto":{"type":"object","properties":{"certified_by_eic":{"type":"boolean","description":"Value of certified_by_eic","example":true},"name_of_eic":{"type":"string","description":"Name of eic certificate","example":"EIC"},"date_of_certification":{"format":"date-time","type":"string","description":"Date of certification"},"invoice_type_id":{"type":"string","description":"Invoice type id","example":"7e82e4fe-6453-4783-a011-81996fd5aea4"},"certified_abstruct_in_stnd_format":{"type":"boolean","description":"Certified abstruct in standard format","example":true},"final_bill":{"type":"boolean","description":"final bill","example":true},"fim_reconcilliation_statement":{"type":"boolean","description":"FIM Reconcilliation Statement","example":true},"no_claim_certification":{"type":"boolean","description":"No claim certification","example":false},"e_way_bill":{"type":"boolean","description":"e_way bill","example":true},"delivery_challan":{"type":"boolean","description":"delivery_Challan","example":true},"lr_copy":{"type":"boolean","description":"lr_copy","example":true},"ld_certification_by_eic":{"type":"boolean","description":"ld_certification_by_eic","example":true},"msme":{"type":"boolean","description":"msme","example":true},"udayam_number":{"type":"string","description":"msme","example":"UDYAM-XY-07-1234567"},"name_of_buyer":{"type":"string","description":"Name of Buyer","example":"Shashi"},"deal_slip":{"type":"string","description":"Deal Slip No","example":"DS30037"},"bill_to_party":{"type":"string","description":"Bill to party","example":"1A2F35846548"},"work_area":{"type":"string","description":"Work Area","example":"Hyderabad"},"vendor_id":{"type":"string","description":"Vendor id","example":"436052fc-8fd1-40bc-aaca-30b715eda7de"},"gst_no":{"type":"string","description":"Gst no","example":"22AAAAA00000A1z5"},"invoice_no":{"type":"string","description":"Invoice No for Vendor Invoice","example":"INV000123"},"inv_date":{"format":"date-time","type":"string","description":"invoice Date"},"inv_basic_value":{"type":"number","description":"Invoice Basic Value!","example":1500},"currency_id":{"type":"string","description":"","example":"8f0356d3-1b2d-4872-8a7f-a979269e0c94"},"gst_type_id":{"type":"string","description":"Vendor gst_type","example":"0908e513-bd16-4224-92f7-83421bfb6fb1"},"cgst_amount":{"type":"number","description":"cgst_amount!","example":135},"sgst_amount":{"type":"number","description":"sgst_amount!","example":135},"igst_amount":{"type":"number","description":"igst_amount!","example":135},"total_invoice_amount":{"type":"number","description":"total_invoice_amount!","example":135},"rendered_service_start_dt":{"format":"date-time","type":"string","description":"rendered_service_start_dt"},"rendered_service_end_dt":{"format":"date-time","type":"string","description":"rendered_service_end_dt"},"job_material_description":{"type":"string","description":"job_material_description","example":"It was populated in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like aldus..."},"attachments":{"description":"List of attachments","example":[{"attachment_type_id":"8a403b87-b64d-4648-a79d-1fbb62c081bb","file_name":"filename","file_path":"www.example.com"}],"type":"array","items":{"$ref":"#/components/schemas/AttachmentDto"}},"invoice_category_id":{"type":"string","description":"invoice_category_id","example":"9d7cf91d-8c5a-41ec-9578-e5d7aaeb363a"}},"required":["invoice_type_id","certified_abstruct_in_stnd_format","final_bill","fim_reconcilliation_statement","no_claim_certification","e_way_bill","delivery_challan","lr_copy","ld_certification_by_eic","msme","udayam_number","name_of_buyer","deal_slip","bill_to_party","work_area","vendor_id","invoice_no","inv_date","inv_basic_value","currency_id","gst_type_id","rendered_service_start_dt","rendered_service_end_dt","job_material_description","attachments","invoice_category_id"]},"InvoiceResponseDto":{"type":"object","properties":{"id":{"type":"string","readOnly":true,"example":"cb9f2c82-2910-4766-a634-54502dfb01e1","format":"v4"},"created_at":{"format":"date-time","type":"string","readOnly":true},"updated_at":{"format":"date-time","type":"string","readOnly":true},"is_active":{"type":"boolean","readOnly":true},"request_no":{"type":"string","readOnly":true,"example":"PR1000001"},"scroll_no":{"type":"string","readOnly":true,"example":"SC-M-DS0002-01"},"invoice_amount":{"type":"number","readOnly":true,"example":100},"invoice_date":{"format":"date-time","type":"string","readOnly":true},"generated_inv_file_path":{"type":"string","readOnly":true,"example":"www.example.com"},"management_approval_dt":{"format":"date-time","type":"string","readOnly":true},"sap_scroll_number":{"type":"string","readOnly":true,"example":"SN1234567"},"sap_scroll_dt":{"format":"date-time","type":"string","readOnly":true},"payment_advice_number":{"type":"string","readOnly":true,"example":"PAN1234567"},"adjusted_amount":{"type":"number","readOnly":true,"example":0},"remaining_amount":{"type":"number","readOnly":true,"example":0},"certified_by_eic":{"type":"boolean","readOnly":true},"name_of_eic":{"type":"string","readOnly":true,"example":"EIC"},"date_of_certification":{"format":"date-time","type":"string","readOnly":true},"certified_abstruct_in_stnd_format":{"type":"boolean","readOnly":true},"final_bill":{"type":"boolean","readOnly":true},"fim_reconcilliation_statement":{"type":"boolean","readOnly":true},"no_claim_certification":{"type":"boolean","readOnly":true},"e_way_bill":{"type":"boolean","readOnly":true},"delivery_challan":{"type":"boolean","readOnly":true},"lr_copy":{"type":"boolean","readOnly":true},"ld_certification_by_eic":{"type":"boolean","readOnly":true},"msme":{"type":"boolean","readOnly":true},"udayam_number":{"type":"string","readOnly":true,"example":"UDYAM-XY-07-1234567"},"invoice_submission_date":{"format":"date-time","type":"string","readOnly":true},"name_of_buyer":{"type":"string","readOnly":true,"example":"Shashi"},"deal_slip":{"type":"string","readOnly":true,"example":"DS30037"},"bill_to_party":{"type":"string","readOnly":true,"example":"1A2F35846548"},"work_area":{"type":"string","readOnly":true,"example":"Hyderabad"},"gst_no":{"type":"string","readOnly":true,"example":"22AAAAA0000A1Z5"},"inv_date":{"format":"date-time","type":"string","readOnly":true},"inv_basic_value":{"type":"number","readOnly":true,"example":1500},"cgst_amount":{"type":"number","readOnly":true,"example":135},"sgst_amount":{"type":"number","readOnly":true,"example":135},"igst_amount":{"type":"number","readOnly":true,"example":135},"total_invoice_amount":{"type":"number","readOnly":true,"example":1770},"rendered_service_start_dt":{"format":"date-time","type":"string","readOnly":true},"rendered_service_end_dt":{"format":"date-time","type":"string","readOnly":true},"job_material_description":{"type":"string","readOnly":true,"example":"It was populated in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like aldus..."},"invoice_no":{"type":"string","readOnly":true,"example":"INV000123"},"project_name":{"type":"string","readOnly":true,"example":"Project Lexys A"},"project_code":{"type":"string","readOnly":true,"example":"100002"},"company_name":{"type":"string","readOnly":true,"example":"Lexys Technologies"},"company_code":{"type":"string","readOnly":true,"example":"12345"},"project_type_code":{"type":"string","readOnly":true,"example":"SDP"},"project_type_des":{"type":"string","readOnly":true,"example":"STEADY PROJECT"},"invoice_type_des":{"type":"string","readOnly":true,"example":"TAX"},"invoice_type_code":{"type":"string","readOnly":true,"example":"TAX"},"invoice_sts_code":{"type":"string","readOnly":true,"example":"STS"},"invoice_sts_des":{"type":"string","readOnly":true,"example":"ST SUBMITTED"},"invoice_category_code":{"type":"string","readOnly":true,"example":"SPPLY"},"invoice_category_des":{"type":"string","readOnly":true,"example":"SUPPLY INVOICE"},"vendor_name":{"type":"string","readOnly":true,"example":"john"},"email":{"type":"string","readOnly":true,"example":"john@email.com"},"is_verified":{"type":"boolean","readOnly":true},"vendor_code":{"type":"string","readOnly":true,"example":"1234"},"currency_code":{"type":"string","readOnly":true,"example":"INR"},"currency_des":{"type":"string","readOnly":true,"example":"INDIAN RUPEES"},"gst_type_code":{"type":"string","readOnly":true,"example":"CGST"}},"required":["id","created_at","updated_at","is_active","request_no","scroll_no","invoice_amount","invoice_date","generated_inv_file_path","management_approval_dt","sap_scroll_number","sap_scroll_dt","payment_advice_number","adjusted_amount","remaining_amount","certified_by_eic","name_of_eic","date_of_certification","certified_abstruct_in_stnd_format","final_bill","fim_reconcilliation_statement","no_claim_certification","e_way_bill","delivery_challan","lr_copy","ld_certification_by_eic","msme","udayam_number","invoice_submission_date","name_of_buyer","deal_slip","bill_to_party","work_area","gst_no","inv_date","inv_basic_value","cgst_amount","sgst_amount","igst_amount","total_invoice_amount","rendered_service_start_dt","rendered_service_end_dt","job_material_description","invoice_no","project_name","project_code","company_name","company_code","project_type_code","project_type_des","invoice_type_des","invoice_type_code","invoice_sts_code","invoice_sts_des","invoice_category_code","invoice_category_des","vendor_name","email","is_verified","vendor_code","currency_code","currency_des","gst_type_code"]},"VendorResponseDto":{"type":"object","properties":{"id":{"type":"string","readOnly":true,"example":"cb9f2c82-2910-4766-a634-54502dfb01e1","format":"v4"},"created_at":{"format":"date-time","type":"string","readOnly":true},"updated_at":{"format":"date-time","type":"string","readOnly":true},"is_active":{"type":"boolean","readOnly":true},"vendor_name":{"type":"string","readOnly":true,"example":"john"},"email":{"type":"string","readOnly":true,"example":"john@email.com"},"role":{"type":"string","readOnly":true,"example":"V"},"vendor_code":{"type":"string","readOnly":true,"example":"34567"},"user_role_des":{"type":"string","readOnly":true,"example":"VENDOR"},"user_role_code":{"type":"string","readOnly":true,"example":"V"},"country_code":{"type":"string","readOnly":true,"example":"IND"},"country_name":{"type":"string","readOnly":true,"example":"INDIA"},"vendor_type_code":{"type":"string","readOnly":true,"example":"IT"},"vendor_type_des":{"type":"string","readOnly":true,"example":"IT"}},"required":["id","created_at","updated_at","is_active","vendor_name","email","role","vendor_code","user_role_des","user_role_code","country_code","country_name","vendor_type_code","vendor_type_des"]},"CreateVendorRequestDto":{"type":"object","properties":{"vendor_name":{"type":"string","description":"vendor_name","example":"john"},"email":{"type":"string","description":"vendor_email","example":"john@email.com"},"user_role_id":{"type":"string","description":"vendor user_role_id","example":"b652b2ca-5732-410c-ac5b-7e157846ad67"},"country_code_id":{"type":"string","description":"vendor country_code","example":"aff32e80-170f-4857-99af-1d2931d7f341"},"vendor_type_id":{"type":"string","example":"5aab488f-0e34-405b-8328-877599733fa3","description":"vendor type_code"},"vendor_code":{"type":"string","example":"1234","description":"vendor_code"}},"required":["vendor_name","email","user_role_id","country_code_id","vendor_type_id","vendor_code"]},"UpdateVendorRequestDto":{"type":"object","properties":{"vendor_name":{"type":"string","description":"vendor_name","example":"john"},"email":{"type":"string","description":"vendor_email","example":"john@email.com"},"user_role_id":{"type":"string","description":"vendor user_role_id","example":"b652b2ca-5732-410c-ac5b-7e157846ad67"},"country_code_id":{"type":"string","description":"vendor_country_code","example":"aff32e80-170f-4857-99af-1d2931d7f341"},"vendor_type_id":{"type":"string","example":"5aab488f-0e34-405b-8328-877599733fa3","description":"vendor type_code"},"vendor_code":{"type":"string","example":"1234","description":"vendor_code"}}},"UserAccessStatusResponseDto":{"type":"object","properties":{"id":{"type":"string","readOnly":true,"example":"cb9f2c82-2910-4766-a634-54502dfb01e1","format":"v4"},"created_at":{"format":"date-time","type":"string","readOnly":true},"updated_at":{"format":"date-time","type":"string","readOnly":true},"is_active":{"type":"boolean","readOnly":true},"invoice_sts_code":{"type":"string","readOnly":true,"example":"STS"},"invoice_sts_des":{"type":"string","readOnly":true,"example":"ST SUBMITTED"},"user_role_des":{"type":"string","readOnly":true,"example":"VENDOR"},"user_role_code":{"type":"string","readOnly":true,"example":"V"}},"required":["id","created_at","updated_at","is_active","invoice_sts_code","invoice_sts_des","user_role_des","user_role_code"]},"UploadRequest":{"type":"object","properties":{"filename":{"type":"string","example":"test.jpg"}},"required":["filename"]},"UploadResponse":{"type":"object","properties":{"url":{"type":"string","readOnly":true,"example":"https://www.example.com"}},"required":["url"]}}},"security":[{"JWTBearerAuth":[]}]}